# Conflicts:
#	coverage/BankDecorator.php.html
#	coverage/BankModel.php.html
#	coverage/BaseModel.php.html
#	coverage/FactoryModel.php.html
#	coverage/FactoryPattent.php.html
#	coverage/OptionUser.php.html
#	coverage/Repository.php.html
#	coverage/UserModel.php.html
#	coverage/dashboard.html
#	coverage/index.html
#	models/UserModel.php
#	phpunit.xml
#	phpunit.xml.bak

+++ models/BankModel.php
@@ -49,9 +49,17 @@ class BankModel extends BaseModel
     // Find user_id trong table banks
     public function findUserByIdTableBank($user_id)
     {
+        if(is_object($user_id)){
+            return false;
+        }
+        elseif(is_array($user_id)){
+            return false;
+        }
+        else{
         $sql = 'SELECT * FROM banks WHERE user_id = ' . $user_id;
         $bank = $this->select($sql);
         return $bank;
+        }
     }
 
     // Get id user new : 
@@ -94,15 +102,20 @@ class BankModel extends BaseModel
         $sql1 = 'SELECT id FROM banks';
         $allUser = $this->select($sql1);
 
-        $_id = $id;
-        $id_start = substr($_id, 3);
-        $id_end = substr($id_start, 0, -3);
+        // $_id = $id;
+        // $id_start = substr($_id, 3);
+        // $id_end = substr($id_start, 0, -3);
 
-        foreach ($allUser as $key) {
-            $md5 = md5($key['id'] . "chuyen-de-web-1");
-            if ($md5 == $id) {
-                $sql = 'DELETE FROM banks WHERE id = ' . $key['id'];
-                return $this->delete($sql);
+        if (is_object($id)) {
+            return false;
+        }
+         else {
+            foreach ($allUser as $key) {
+                $md5 = md5($key['id'] . "chuyen-de-web-1");
+                if ($md5 == $id) {
+                    $sql = 'DELETE FROM banks WHERE id = ' . $key['id'];
+                    return $this->delete($sql);
+                }
             }
         }
     }
